HIGHLIGHTS/requirements.txt
stable-baselines3[extra]>=1.6.0
rl-baselines3-zoo>=1.6.0
huggingface-sb3>=2.2.0
av==10.0.0
cloudpickle==1.6.0
cycler==0.11.0
fonttools==4.38.0
gym>=0.21.0
gym-notices==0.0.8
imageio==2.22.2
importlib-metadata==5.0.0
kiwisolver==1.4.4
matplotlib==3.5.3
numpy>=1.21.6
opencv-python==4.6.0.66
packaging==21.3
pandas>=1.3.5
Pillow==9.2.0
pygame==2.1.2
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.5
scipy==1.7.3
six==1.16.0
torch>=1.12.1
typing_extensions==4.4.0
zipp==3.10.0
pyyaml>=5.4.1
tqdm>=4.62.0
tensorboard>=2.8.0
shimmy>=0.2.1
ale-py>=0.7.4
